Fix prometheus startup and metrics folders
