+++ .gitignore
@@ -2,3 +2,4 @@ log_bridge.log
 __pycache__
 metrics
 .env
+prometheus.yml
