+++ argo_bridge.py
@@ -8,48 +8,8 @@ import logging
 import argparse
 from flask_cors import CORS  # Add this import
 import httpx
-from prometheus_client import Counter, Histogram, generate_latest, CONTENT_TYPE_LATEST, multiprocess, CollectorRegistry
 from functools import wraps
 
-# This ensures metrics work correctly with multiple Gunicorn workers
-os.environ.setdefault('PROMETHEUS_MULTIPROC_DIR', 'metrics')
-
-# Initialize Prometheus metrics - just the essentials
-REQUEST_COUNT = Counter('argo_requests_total', 'Total API requests', ['endpoint', 'model', 'status'])
-REQUEST_LATENCY = Histogram('argo_request_latency_seconds', 'API request latency', ['endpoint', 'model'])
-
-def track_metrics(endpoint_name):
-    """Decorator to track metrics for API endpoints"""
-    def decorator(func):
-        @wraps(func)
-        def wrapper(*args, **kwargs):
-            # Extract model from request
-            data = request.get_json(silent=True) or {}
-            model = data.get('model', DEFAULT_MODEL) if data else None
-            
-            # Start timing
-            start_time = time.time()
-            
-            try:
-                # Execute the endpoint function
-                response = func(*args, **kwargs)
-                
-                # Record metrics after successful execution
-                status = response[1] if isinstance(response, tuple) else response.status_code if hasattr(response, 'status_code') else 200
-                REQUEST_COUNT.labels(endpoint=endpoint_name, model=model, status=status).inc()
-                REQUEST_LATENCY.labels(endpoint=endpoint_name, model=model).observe(time.time() - start_time)
-                print('collected data', flush=True)
-                
-                return response
-            except Exception as e:
-                # Record failure metrics
-                REQUEST_COUNT.labels(endpoint=endpoint_name, model=model, status=500).inc()
-                REQUEST_LATENCY.labels(endpoint=endpoint_name, model=model).observe(time.time() - start_time)
-                # Re-raise the exception
-                raise e
-                
-        return wrapper
-    return decorator
 
 app = Flask(__name__)
 CORS(app, 
@@ -205,7 +165,6 @@ def get_api_url(model, endpoint_type):
 @app.route('/chat/completions', methods=['POST'])
 @app.route('/api/chat/completions', methods=['POST'])
 @app.route('/v1/chat/completions', methods=['POST']) #LMStudio Compatibility
-@track_metrics('chat_completions')
 def chat_completions():
     logging.info("Received chat completions request")
 
@@ -239,6 +198,7 @@ def chat_completions():
         return Response(_stream_chat_response(model, req_obj), mimetype='text/event-stream')
     else:
         response = requests.post(get_api_url(model, 'chat'), json=req_obj)
+        
         if not response.ok:
             logging.error(f"Internal API error: {response.status_code} {response.reason}")
             return jsonify({"error": {
@@ -330,7 +290,6 @@ def _static_chat_response(text, model):
 
 @app.route('/completions', methods=['POST'])
 @app.route('/v1/completions', methods=['POST', 'OPTIONS']) #LMStudio Compatibility
-@track_metrics('completions')
 def completions():
     logging.info("Received completions request")
     data = request.get_json()
@@ -414,7 +373,6 @@ def _stream_completions_response(text, model):
 """
 @app.route('/embeddings', methods=['POST'])
 @app.route('/v1/embeddings', methods=['POST'])
-@track_metrics('embeddings')
 def embeddings():
     logging.info("Recieved embeddings request")
     data = request.get_json()
@@ -487,7 +445,6 @@ def _get_embeddings_from_argo(texts, model):
 """
 @app.route('/models', methods=['GET'])
 @app.route('/v1/models', methods=['GET'])
-@track_metrics('models')
 def models_list():
     logging.info("Received models list request")
     return jsonify(MODELS)
@@ -532,25 +489,6 @@ def check_argo_connection():
 
 """
 =================================
-    Metrics Endpoint
-=================================
-"""
-
-@app.route('/metrics', methods=['GET'])
-def metrics():
-    metrics_token = os.environ.get('METRICS_TOKEN', 'default_secret_token')
-    
-    # Check for auth token in header
-    auth_header = request.headers.get('Authorization')
-    if not auth_header or auth_header != f'Bearer {metrics_token}':
-        return Response("Access denied: Invalid or missing token", status=403)
-
-    registry = CollectorRegistry()
-    multiprocess.MultiProcessCollector(registry)
-    return Response(generate_latest(registry), mimetype=CONTENT_TYPE_LATEST)
-
-"""
-=================================
     CLI Functions
 =================================
 """
@@ -565,7 +503,6 @@ def parse_args():
 
 if __name__ == '__main__':
     args = parse_args()
-    
     debug_enabled = args.dlog
     logging.basicConfig(
         filename=ANL_DEBUG_FP,
