+++ docker-compose.yaml
@@ -4,8 +4,11 @@ services:
     ports:
       - "80:80"
     restart: unless-stopped
+    volumes:
+      - metrics_data:/app/metrics
     environment:
       - METRICS_TOKEN=${METRICS_TOKEN:-default_secret_token}
+      - PROMETHEUS_MULTIPROC_DIR=/app/metrics
 
   prometheus:
     image: prom/prometheus:latest
@@ -36,7 +39,7 @@ services:
       - "127.0.0.1:3000:3000"
     restart: unless-stopped
     environment:
-      - GF_SECURITY_ADMIN_PASSWORD=admin_password  # Change this!
+      - GF_SECURITY_ADMIN_PASSWORD=admin_password  
       - GF_USERS_ALLOW_SIGN_UP=false
       - GF_AUTH_ANONYMOUS_ENABLED=false
       - GF_SERVER_ROOT_URL=http://localhost:3000
@@ -46,3 +49,4 @@ services:
 volumes:
   prometheus_data:
   grafana_data:
+  metrics_data:
